requirements.txt
chainlit
google-generativeai
python-dotenv
langchain
langchain-google-genai
chromadb
pdfplumber
python-docx